requirements.txt
MetaTrader5
pandas
numpy
fastapi
uvicorn
websockets
python-multipart
scikit-learn
torch
transformers
xgboost
lightgbm
ta-lib
plotly
dash
requests
python-dotenv
google-generativeai
onnxruntime-directml
